Massive update - reorganization, consistency, file writing, finalized error handling
