+++ calculator.py
@@ -7,150 +7,204 @@ import numpy as np
 from .utils import cleaner
 from .message import warning,log,display
 
+#######################################################################################################
+#Calculator object
 
 class Calculator:
-    ''' calculator object is created and serves as the IO interface
-    
-    #create the mol object
-    mol = SmilesToMol('CCC')
-
-    #create the calculator object by calling the setup for
-    #whatever program is actually run
-    gaussian_calc = Gaussian.opt(mol,params)
-
-    #the the calculator object manages the writing files and submission
-    result_mol = gaussian_calc.run()
-
-    #that way I can do a batch submission 
-
-
-    calculators = [Gauss.opt(conf) for conf in mol.conformers]
-    RunParallel(calculators)
-    
-    to run them as an array job
-
-
-    so need a 'Run','RunParallel', 'RunSerial' function to
-    take in a calculator and write an appropriate sbatch
-
-    
-    so what does this object actually consist of?
-    i think its mostly just a container, and then writes the files
-
-
-
-    it should take in a string that is the input file,
-    it should know the number of cores/memory/slurm stuff
-    it should know the command to run
-    if should know the program called
-
-    '''
-
+    '''Calculator object is the middle-person between the actual 
+    program and running the calculations'''
     
     def __init__(self,
-                 input=None,
-                 command=None,
-                 nproc=None,
-                 mem=None,
-                 time=None,
-                 partition=None,
-                 program=None,
-                 runtype=None,
-                 jobname=None,
-                 mol=None):
-
-
-        self.input = input
-        self.command = command
+                jobname,
+                input,
+                command,
+                nproc,
+                mem,
+                time,
+                partition,
+                program,
+                mol,
+                argument_dict,
+                try_count):
+        
+    ##############
+    #slurm inputs#
         self.nproc = nproc
         self.mem = mem
         self.time = time
         self.partition = partition
         self.program = program
-        self.runtype = runtype
         self.jobname = jobname
-        self.dir = os.path.abspath(jobname) + '/'
-        self.inputfile_full = self.dir +  self.jobname + '.' + self.program.infiles[0]
-        self.outputfile_full = self.dir +  self.jobname + '.' + self.program.outfiles[0]
-        self.inputfile_relative = self.jobname + '.' + self.program.infiles[0]
-        self.outputfile_relative = self.jobname + '.' + self.program.outfiles[0]
-        self.mol = mol
 
+    ##############
+    #resubmission#
+        self.try_count = try_count
+        self.argument_dict = argument_dict
+
+    #########
+    #file io#
+        self.dir = os.path.abspath(jobname) + '/'
+        self.basename_full = self.dir +  self.jobname + '-try{0}'.format(try_count) 
+        self.inputfile_full = self.dir +  self.jobname + '-try{0}'.format(try_count) + '.' + self.program.infiles[0]
+        self.outputfile_full = self.dir +  self.jobname + '-try{0}'.format(try_count) + '.' + self.program.outfiles[0]
+        self.inputfile_relative = self.jobname + '-try{0}'.format(try_count) + '.' + self.program.infiles[0]
+        self.outputfile_relative = self.jobname + '-try{0}'.format(try_count) + '.' + self.program.outfiles[0]
+
+    ############
+    #input file#
+        self.input = input
         #write the input file in the right directory
         #creating the directory if it does not exist
         if not os.path.isdir(self.dir):
             os.makedirs(self.dir)
-        
+
         for index,content in enumerate(self.input):
-            with open('{0}/{1}.{2}'.format(self.dir,self.jobname,self.program.infiles[index]),'w') as inputfile:
+            with open('{0}/{1}-try{2}.{3}'.format(self.dir,self.jobname,self.try_count,self.program.infiles[index]),'w') as inputfile:
                 inputfile.write(content)
 
+    ############
+    #submission#
+        self.command = command
         #replace the generic command with the input/output files
         self.command = self.command.replace('INPUTFILE',self.inputfile_relative)
         self.command = self.command.replace('OUTPUTFILE',self.outputfile_relative)
 
-        def InputFile(self):
-            display(self.input)
-        def Command(self):
-            display(self.command)
+        #always return a modified copy of the mol, rather than the modifying the mol itself
+        self.mol = mol.copy()
+
+    #########
+    #methods#
+
+    #can be called by the Run or RunBatch functions to update the calculator
+    #object based on the failures read from the current submission
+    def resubmit(self):
+        updated_argument_dict = self.program.fix_errors(self.mol,self.basename_full,self.argument_dict)
+        updated_calculator = self.program.resubmit(updated_argument_dict,try_count=self.try_count+1)
+        return(updated_calculator)
+
+    #miscilaneous helper functions
+    def InputFile(self):
+        display(self.input)
+    def Command(self):
+        display(self.command)
 
 
-def Run(calculator):
+
+#######################################################################################################
+#Running the calculators
+
+def Run(calculator,tries=1):
     '''just submit a single job'''
+
+#######################
+#check for valid input#
     #check that the input is a calculator object
     if not isinstance(calculator,Calculator):
         raise TypeError('Expected Calcluator object')
-
-    sbatch_name = '{0}.sbatch'.format(calculator.jobname)
-    with open('{0}{1}'.format(calculator.dir,sbatch_name),'w') as sbatch: #directory is already ended with /
-        sbatch.write(generic_single(calculator))
-    
-    #submit the job
-    log('submitting - sbatch {0}'.format(sbatch_name))
-    p = subprocess.Popen('sbatch {0}'.format(sbatch_name), stdout=subprocess.PIPE, shell=True,cwd=calculator.dir)
-    output, err = p.communicate()
-    p_status = p.wait()
-    slurmID = str(output).strip().split()[-1][:-3]
-    slurmoutput = '{0}{1}-{2}.out'.format(calculator.dir,calculator.jobname,slurmID) #directory is already ended with /
-
-    return(calculator.program.read_output(calculator,slurmoutput))
     
 
-def RunBatch(calculators,jobname='batch_job',max=50):
+##############################################
+#start submissions while warnings are present#
+    calculator.mol.warnings = ['First_submission']
+
+    #submit as many times as is requested until the job finishes without warnings
+    while len(calculator.mol.warnings) > 0:
+
+        #if it's not the first submission, update the calculator 
+        #with the resubmisson method
+        if not calculator.mol.warnings[0] == 'First_submission':
+            calculator = calculator.resubmit()
+
+        if calculator.try_count <= tries:
+            sbatch_name = '{0}-try{1}.sbatch'.format(calculator.jobname,calculator.try_count)
+            with open('{0}{1}'.format(calculator.dir,sbatch_name),'w') as sbatch: #directory is already ended with /
+                sbatch.write(generic_single(calculator))
+
+            #submit the job
+            log('submitting - sbatch {0}'.format(sbatch_name))
+            p = subprocess.Popen('sbatch {0}'.format(sbatch_name), stdout=subprocess.PIPE, shell=True,cwd=calculator.dir)
+            output, err = p.communicate()
+            p_status = p.wait()
+            slurmID = str(output).strip().split()[-1][:-3]
+            slurmoutput = '{0}{1}-{2}.out'.format(calculator.dir,calculator.jobname,slurmID) #directory is already ended with /
+
+            calculator.mol = calculator.program.read_output(calculator,slurmoutput)
+
+        else:
+            raise IndexError('Ran out of resubmission tries')
+
+    return(calculator.mol)
+
+
+
+def RunBatch(calculators,jobname='batch_job',max=50,tries=1):
     '''Take a list of calculators and run all of them as a slurm array - assume they all have the same resources as the first one'''
+    
+#######################
+#check for valid input#
     #check that the input is a calculator object
     for calculator in calculators:
         if not isinstance(calculator,Calculator):
-            raise TypeError('Expected list of Calcluator objects')
-    
+            raise TypeError('Expected iterable of Calcluator objects')
+
+#####################
+#prep for submission#
     #create the directory to keep all of the batch job scripts
     if not os.path.isdir(jobname):
         os.makedirs(jobname)
-    
-    with open(jobname + '/' + jobname+ '.sbatch','w') as sbatch:
-        sbatch.write(generic_batch(jobname,calculators,max))
-    for index,calculator in enumerate(calculators):
-        command_file_name = '{0}/{0}-{1}.sh'.format(jobname,index+1)   #slurm array index start at 1 not 0
-        with open(command_file_name,'w') as command_file:
-            command_file.write(generic_command(calculator))
-    
-    #submit the job
-    log('submitting - sbatch {0}'.format(jobname))
-    p = subprocess.Popen('sbatch {0}.sbatch'.format(jobname), stdout=subprocess.PIPE, shell=True,cwd=jobname)
-    output, err = p.communicate()
-    p_status = p.wait()
-    slurmID = str(output).strip().split()[-1][:-3]
 
-    output = []
-    for index,calculator in enumerate(calculators):
-        slurmoutput = '{0}/{0}-{1}_{2}.out'.format(jobname,slurmID,index+1) #directory is already ended with /
-        output.append(calculator.program.read_output(calculator,slurmoutput))
+    output = [None for calculator in calculators]
+    #these are the indicies of the calculators that need to be submitted
+    need_resub = list(range(0,len(calculators)))
+
+##############################################
+#start submissions while warnings are present#
+    #while there are jobs with warnings, keep resubmitting
+    while len(need_resub) > 0:
+
+        #unless the number of tries is up (just need to check the first calculator in the list that needs computing)
+        if need_resub[0].try_count <= tries:
+
+            with open(jobname + '/' + jobname+ '.sbatch','w') as sbatch:
+                sbatch.write(generic_batch(jobname,calculators,max,need_resub))
+
+            for index in enumerate(need_resub):
+                command_file_name = '{0}/{0}-{1}-try{2}.sh'.format(jobname,index+1,calculators[index].try_count)   #slurm array index start at 1 not 0
+                with open(command_file_name,'w') as command_file:
+                    command_file.write(generic_command(calculator))
+        
+            #submit the job
+            log('submitting - sbatch {0}'.format(jobname))
+            p = subprocess.Popen('sbatch {0}.sbatch'.format(jobname), stdout=subprocess.PIPE, shell=True,cwd=jobname)
+            output, err = p.communicate()
+            p_status = p.wait()
+            slurmID = str(output).strip().split()[-1][:-3]
+
+        
+            #gather the output and account for failures
+            new_need_resub = []
+            for index in enumerate(need_resub):
+                slurmoutput = '{0}/{0}-{1}_{2}.out'.format(jobname,slurmID,index+1) 
+                calculators[index].mol = calculator[index].program.read_output(calculators[index],slurmoutput)
+                output[index] = calculators[index].mol
+
+                #if the returned mol object has warnings, 
+                #use the calculator.resubmit() method to update the calculator
+                if len(calculators[index].mol.warnings) > 0:
+                    new_need_resub.append(index)
+                    calculators[index] = calculators[index].resubmit()
+
+            need_resub = new_need_resub
+
+        else:
+            raise IndexError('Ran out of resubmission tries')
+    
     return(output)
 
 
+
 #############################################################################################################################
-#templates
-#############################################################################################################################
+#submission templates
 
 def generic_single(calculator):
     generic_single_string = r"""#!/bin/bash
@@ -169,7 +223,7 @@ def generic_single(calculator):
 """.format(calculator.jobname,calculator.partition,calculator.mem,calculator.nproc,calculator.time,calculator.command)
     return(generic_single_string)
 
-def generic_batch(jobname,calculators,max):
+def generic_batch(jobname,calculators,max,need_resub):
     #need to get the slurm parameters - get it from the first calculator and just assume they are the same....
     partition = calculators[0].partition
     mem = calculators[0].mem
@@ -193,7 +247,7 @@ chmod 777 {0}-${{SLURM_ARRAY_TASK_ID}}.sh
 
 ./{0}-${{SLURM_ARRAY_TASK_ID}}.sh
 
-""".format(jobname,partition,mem,nproc,time,len(calculators),max)
+""".format(jobname,partition,mem,nproc,time,len(need_resub),max)
     return(generic_batch_string)
 
 def generic_command(calculator):
